Docker image for development
Adds a Dockerfile which can be used to build both development and production
versions of the website.

Building the image and running the container is handled in `docker_build.sh`
and `docker_run.sh` respectively. By default, the image and container will be
built and run with the development configuration. If you want to test a
production version of the image, you can pass `production` as the only
argument to both `docker_build.sh` and `docker_run.sh`.

To prevent overwrites of the `node_modules` folder, only required folders in
the root directory are mounted in the development container. To reduce the
number of bind mounts required, the project source has been moved into the
`src` folder.